3.5 second   commit

+++ backend/config/cloudinary.js
@@ -1,6 +1,6 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
-const uploadOnCloudinary = async () => {
+const uploadOnCloudinary = async (filePath) => {
   cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
@@ -9,11 +9,15 @@ const uploadOnCloudinary = async () => {
 
   try {
     const uploadResult = await cloudinary.uploader.upload(filePath);
-    fs.unlinkSync(filePath);
+    if (fs.existsSync(filePath)) {
+      fs.unlinkSync(filePath);
+    }
     return uploadResult.secure_url;
   } catch (error) {
-    fs.unlinkSync(filePath);
-    return res.status(500).json({ message: "Error uploading file to Cloudinary" });
+    if (fs.existsSync(filePath)) {
+      fs.unlinkSync(filePath);
+    }
+    throw error;
   }
 };
